% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cpas_get}
\alias{print.cpas_get}
\title{Print Method for cpas_get Class}
\usage{
\method{print}{cpas_get}(x, ...)
}
\arguments{
\item{x}{An object of class cpas_get.}

\item{...}{Additional arguments passed to print.}
}
\value{
No return value, prints a summary of the API response.
}
\description{
Custom print method for the result object returned by get_data function.
}
\examples{
\dontrun{
   ## The raw API accessor, printed.
   r <- get_data("GSE14814", "surv_data")
   print(r)
}
}
