% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cpas_get_expr}
\alias{print.cpas_get_expr}
\title{Print Method for cpas_get_expr Class}
\usage{
\method{print}{cpas_get_expr}(x, ...)
}
\arguments{
\item{x}{An object of class cpas_get_expr.}

\item{...}{Additional arguments passed to print.}
}
\value{
No return value, prints a summary of the gene expression data retrieval.
}
\description{
Custom print method for the result object returned by get_expr_data function.
}
\examples{
\dontrun{
   e <- get_expr_data("GSE14814", c("GAPDH", "ACTB"), process_duplicates = "max")
   print(e)
}
}
