% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cpas_merge}
\alias{print.cpas_merge}
\title{Print Method for cpas_merge Class}
\usage{
\method{print}{cpas_merge}(x, ...)
}
\arguments{
\item{x}{An object of class cpas_merge.}

\item{...}{Additional arguments passed to print.}
}
\value{
No return value, prints a summary of the merged data.
}
\description{
Custom print method for the result object returned by merge_surv_expr function.
}
\examples{
\dontrun{
   e <- get_expr_data("GSE14814", c("GAPDH", "ACTB"), process_duplicates = "max")
   m <- merge_surv_expr("GSE14814", e)
   print(m)
}
}
