% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_methods.R
\name{print.cpas_signature}
\alias{print.cpas_signature}
\title{Print Method for cpas_signature Class}
\usage{
\method{print}{cpas_signature}(x, ...)
}
\arguments{
\item{x}{An object of class cpas_signature.}

\item{...}{Additional arguments passed to print.}
}
\value{
No return value, prints a summary of the gene signature calculation.
}
\description{
Custom print method for the result object returned by get_signature_value function.
}
\examples{
\dontrun{
   s <- get_signature_value("0.5*GAPDH + 0.5*ACTB", "GSE14814")
   print(s)
}
}
