% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpas_meta.R
\name{plot_meta_forest}
\alias{plot_meta_forest}
\title{整合分析森林图}
\usage{
plot_meta_forest(x, digits = 4, ...)
}
\arguments{
\item{x}{一个 \code{cpas_meta} 对象。}

\item{digits}{图中标注保留的小数位(默认 4)。}

\item{...}{保留，暂未使用。}
}
\value{
ggplot 对象。
}
\description{
绘制 \code{\link{cpas_meta}} 的逐队列 HR(95\% CI) 与合并 HR 森林图。
}
\examples{
\dontrun{
   m <- cpas_meta(c("GSE31210", "GSE37745"), marker = "TP53", type = "DFS")
   plot_meta_forest(m, digits = 4)
}
}
